+++ jest.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  clearMocks: true,
+  modulePathIgnorePatterns: ['<rootDir>/dist/', '<rootDir>/example/'],
+  preset: 'ts-jest',
+  testEnvironment: 'jsdom',
+};