+++ src/api/SignalClient.ts
@@ -26,6 +26,7 @@ import {
   SubscribedQualityUpdate,
   SubscriptionPermission,
   SubscriptionPermissionUpdate,
+  SubscriptionResponse,
   SyncState,
   TrackPermission,
   TrackPublishedResponse,
@@ -134,6 +135,8 @@ export class SignalClient {
 
   onSubscriptionPermissionUpdate?: (update: SubscriptionPermissionUpdate) => void;
 
+  onSubscriptionError?: (update: SubscriptionResponse) => void;
+
   onLocalTrackUnpublished?: (res: TrackUnpublishedResponse) => void;
 
   onTokenRefresh?: (token: string) => void;
@@ -221,7 +224,7 @@ export class SignalClient {
 
     return new Promise<JoinResponse | ReconnectResponse | void>(async (resolve, reject) => {
       const abortHandler = async () => {
-        await this.close();
+        this.close();
         reject(new ConnectionError('room connection has been cancelled (signal)'));
       };
 
@@ -597,6 +600,10 @@ export class SignalClient {
     } else if (msg.case === 'trackUnpublished') {
       if (this.onLocalTrackUnpublished) {
         this.onLocalTrackUnpublished(msg.value);
+      }
+    } else if (msg.case === 'subscriptionResponse') {
+      if (this.onSubscriptionError) {
+        this.onSubscriptionError(msg.value);
       }
     } else if (msg.case === 'pong') {
       this.resetPingTimeout();