+++ CHANGELOG.md
@@ -1,5 +1,77 @@
 # Change Log
 
+## 1.11.1
+
+### Patch Changes
+
+- Fix mute event handling - [#740](https://github.com/livekit/client-sdk-js/pull/740) ([@davidzhao](https://github.com/davidzhao))
+
+## 1.11.0
+
+### Minor Changes
+
+- Increase default adaptiveStream pixelDensity on high-density(mobile) screens - [#735](https://github.com/livekit/client-sdk-js/pull/735) ([@davidzhao](https://github.com/davidzhao))
+
+- Replace event emitter lib with eventemitter3 - [#681](https://github.com/livekit/client-sdk-js/pull/681) ([@lukasIO](https://github.com/lukasIO))
+
+### Patch Changes
+
+- Handle device mute and document freeze events - [#734](https://github.com/livekit/client-sdk-js/pull/734) ([@davidzhao](https://github.com/davidzhao))
+
+- Pass method logLevel to LogExtension instead of configured logLevel - [#730](https://github.com/livekit/client-sdk-js/pull/730) ([@lukasIO](https://github.com/lukasIO))
+
+- Fix svc encodings for safari and chrome before 113 - [#731](https://github.com/livekit/client-sdk-js/pull/731) ([@cnderrauber](https://github.com/cnderrauber))
+
+- Always catch reconnectFuture rejections - [#727](https://github.com/livekit/client-sdk-js/pull/727) ([@HermanBilous](https://github.com/HermanBilous))
+
+- Work around iOS safari audio playback issue when not publishing, by playing back silent audio - [#733](https://github.com/livekit/client-sdk-js/pull/733) ([@lukasIO](https://github.com/lukasIO))
+
+## 1.10.0
+
+### Minor Changes
+
+- Add track processor API - [#711](https://github.com/livekit/client-sdk-js/pull/711) ([@lukasIO](https://github.com/lukasIO))
+
+### Patch Changes
+
+- Use replaceTrack(null) for pauseUpstream - [#716](https://github.com/livekit/client-sdk-js/pull/716) ([@davidzhao](https://github.com/davidzhao))
+
+- Always add codec info to AddTrackRequest - [#728](https://github.com/livekit/client-sdk-js/pull/728) ([@cnderrauber](https://github.com/cnderrauber))
+
+- Surface subscription error via TrackEvent.SubscriptionFailed when trying to subsribe to an unsupported codec - [#722](https://github.com/livekit/client-sdk-js/pull/722) ([@lukasIO](https://github.com/lukasIO))
+
+- Reject signal connection promise immediately when aborted - [#719](https://github.com/livekit/client-sdk-js/pull/719) ([@lukasIO](https://github.com/lukasIO))
+
+- Fix svc mode for chrome v113 - [#720](https://github.com/livekit/client-sdk-js/pull/720) ([@cnderrauber](https://github.com/cnderrauber))
+
+## 1.9.7
+
+### Patch Changes
+
+- Fix browser parser check - [#714](https://github.com/livekit/client-sdk-js/pull/714) ([@lukasIO](https://github.com/lukasIO))
+
+- Ensure same framerates for iOS RN simulcast - [#710](https://github.com/livekit/client-sdk-js/pull/710) ([@davidliu](https://github.com/davidliu))
+
+## 1.9.6
+
+### Patch Changes
+
+- Make sure `TrackUnsubscribed` events are emitted before the publication gets deleted from maps - [#708](https://github.com/livekit/client-sdk-js/pull/708) ([@lukasIO](https://github.com/lukasIO))
+
+- Use body instead of document as intersection observer root - [#703](https://github.com/livekit/client-sdk-js/pull/703) ([@lukasIO](https://github.com/lukasIO))
+
+- Use default video dimensions when they are not available - [#709](https://github.com/livekit/client-sdk-js/pull/709) ([@davidzhao](https://github.com/davidzhao))
+
+## 1.9.5
+
+### Patch Changes
+
+- Remove ua-parser-js dependency and fix browser version comparison - [#697](https://github.com/livekit/client-sdk-js/pull/697) ([@lukasIO](https://github.com/lukasIO))
+
+- Use STATE_MISMATCH disconnect reason in connection reconciliation - [#705](https://github.com/livekit/client-sdk-js/pull/705) ([@lukasIO](https://github.com/lukasIO))
+
+- Make sure engine gets closed when connection reconciliation triggers - [#702](https://github.com/livekit/client-sdk-js/pull/702) ([@lukasIO](https://github.com/lukasIO))
+
 ## 1.9.4
 
 ### Patch Changes